Move everything into ShortestPath\SingleSource
The new algorithms now use a smaller signature and avoid passing the
start Vertex to the ctor.
Keep a BC layer for now.
